Move the AssetManager to Graphics
